Move writers to new io package
