+++ invisible_cities/io/dst_io_test.py
@@ -5,9 +5,9 @@ import tables as tb
 from pytest import mark
 
 from .. core.test_utils    import assert_dataframes_equal
-from .. reco.dst_io        import kr_writer
-from .. reco.dst_io        import KrEvent
 from .. reco.dst_functions import load_dst
+from .  dst_io             import kr_writer
+from .  dst_io             import KrEvent
 
 
 def test_Kr_writer(config_tmpdir, Kr_dst_data):
