+++ invisible_cities/io/run_and_event_io.py
@@ -1,5 +1,5 @@
-from . import nh5           as table_formats
-from . import tbl_functions as tbl
+from ..reco import nh5           as table_formats
+from ..reco import tbl_functions as tbl
 
 def _make_run_event_tables(hdf5_file, compression):
